+++ clouddriver-google/src/main/groovy/com/netflix/spinnaker/clouddriver/google/batch/GoogleBatchRequest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.netflix.spinnaker.clouddriver.googlecommon.batch;
+package com.netflix.spinnaker.clouddriver.google.batch;
 
 import com.google.api.client.googleapis.batch.BatchRequest;
 import com.google.api.client.googleapis.batch.json.JsonBatchCallback;
